+++ ip/imgloader.py
@@ -23,6 +23,9 @@ class ImageLoader(metaclass=abc.ABCMeta):
       return ImageLoaderDefault(img_filename)
     raise AttributeError('image_loader_factory: ' + pixel_format + ' not found')
 
+  def get_native_channels(self):
+    pass
+
 
 class ImageLoaderDefault(ImageLoader):
   def __init__(self, filename):
@@ -30,6 +33,9 @@ class ImageLoaderDefault(ImageLoader):
 
   def imread(self):
     return cv2.imread(self.__filename)
+
+  def get_native_channels(self):
+    return self.imread()
 
 
 class ImageLoaderRawNV21(ImageLoader):
@@ -49,12 +55,19 @@ class ImageLoaderRawNV21(ImageLoader):
     raw_img = self._read_raw()
     return cv2.cvtColor(raw_img, cv2.COLOR_YUV2BGR_NV21)
 
+  def get_native_channels(self):
+    bgr_converted_yuv = self.imread()
+    return  cv2.cvtColor(bgr_converted_yuv, cv2.COLOR_BGR2YUV)
+
 
 class ImageLoaderRawNV12(ImageLoaderRawNV21):
   def imread(self):
     raw_img = self._read_raw()
     return cv2.cvtColor(raw_img, cv2.COLOR_YUV2BGR_NV12)
 
+  def get_native_channels(self):
+    bgr_converted_yuv = self.imread()
+    return  cv2.cvtColor(bgr_converted_yuv, cv2.COLOR_BGR2YUV)
 
 class ImageLoaderRawI420(ImageLoaderRawNV21):
   def imread(self):